remove duplicatd code & translate print UI
