+++ frontend/src/components/print/print-nuxt/DownloadNuxtPdfButton.vue
@@ -2,8 +2,9 @@
   <div>
     <v-btn color="primary" :loading="loading" outlined @click="generatePdf">
       <v-icon>mdi-printer</v-icon>
-      <div class="mx-1">with</div>
-      <v-icon>mdi-nuxt</v-icon>
+      <div class="mx-1">
+        {{ $tc('components.print.printNuxt.downloadNuxtPdfListItem.label') }}
+      </div>
     </v-btn>
     <v-snackbar v-model="error" :timeout="10000">
       {{ $tc('components.print.localPdfDownloadButton.error') }}
@@ -17,61 +18,10 @@
 </template>
 
 <script>
-import { saveAs } from 'file-saver'
-import slugify from 'slugify'
-
-import axios from 'axios'
-
-const PRINT_SERVER = window.environment.PRINT_SERVER
+import { generatePdfMixin } from './generatePdfMixin.js'
 
 export default {
   name: 'DownloadNuxtPdfButton',
-  props: {
-    config: {
-      type: Object,
-      default: () => {},
-    },
-  },
-  data() {
-    return {
-      loading: false,
-      error: null,
-    }
-  },
-  methods: {
-    async generatePdf() {
-      this.loading = true
-
-      try {
-        const response = await axios({
-          method: 'get',
-          url: `${PRINT_SERVER}/server/pdfChrome?config=${encodeURIComponent(
-            JSON.stringify(this.config)
-          )}`,
-          responseType: 'arraybuffer',
-          withCredentials: true,
-          headers: {
-            'Cache-Control': 'no-cache',
-            Pragma: 'no-cache',
-            Expires: '0',
-          },
-        })
-
-        saveAs(
-          new Blob([response.data]),
-          slugify(this.config.documentName, {
-            locale: this.$store.state.lang.language.substr(0, 2),
-          })
-        )
-      } catch (error) {
-        this.error = error
-        console.log(error)
-      } finally {
-        this.loading = false
-      }
-    },
-  },
+  mixins: [generatePdfMixin],
 }
 </script>
-
-<style scoped></style>
