+++ frontend/src/components/print/print-react/DownloadReactPdfButton.vue
@@ -1,9 +1,10 @@
 <template>
   <div>
-    <v-btn color="primary" :loading="loading" outlined @click="generatePdf">
+    <v-btn color="primary" :loading="loading" outlined @click.stop="generatePdf">
       <v-icon>mdi-printer</v-icon>
-      <div class="mx-1">with</div>
-      <v-icon>mdi-react</v-icon>
+      <div class="mx-1">
+        {{ $tc('components.print.printReact.downloadNuxtPdfListItem.label') }}
+      </div>
     </v-btn>
     <v-snackbar v-model="error" :timeout="10000">
       {{ $tc('components.print.localPdfDownloadButton.error') }}
@@ -17,56 +18,10 @@
 </template>
 
 <script>
-import { saveAs } from 'file-saver'
-import slugify from 'slugify'
-
-const RENDER_IN_WORKER = true
+import { generatePdfMixin } from './generatePdfMixin.js'
 
 export default {
   name: 'DownloadReactPdfButton',
-  props: {
-    config: {
-      type: Object,
-      default: () => {},
-    },
-  },
-  data() {
-    return {
-      loading: false,
-      error: null,
-    }
-  },
-  methods: {
-    async generatePdf() {
-      this.loading = true
-
-      // lazy load generatePdf to avoid loading complete react-pdf when showing PDF download button
-      const generatePdfModule = await import('./generatePdf.js')
-
-      const { blob, error } = await generatePdfModule.generatePdf({
-        config: { ...this.config, apiGet: this.api.get.bind(this) },
-        storeData: this.$store.state,
-        translationData: this.$i18n.messages,
-        renderInWorker: RENDER_IN_WORKER,
-      })
-
-      this.loading = false
-
-      if (error) {
-        this.error = error
-        console.log(error)
-        return
-      }
-
-      saveAs(
-        blob,
-        slugify(this.config.documentName, {
-          locale: this.$store.state.lang.language.substr(0, 2),
-        })
-      )
-    },
-  },
+  mixins: [generatePdfMixin],
 }
 </script>
-
-<style scoped></style>
