+++ frontend/src/components/print/print-react/DownloadReactPdfListItem.vue
@@ -2,7 +2,7 @@
   <v-list-item :disabled="loading" @click="generatePdf">
     <v-list-item-icon>
       <v-icon v-if="loading" class="mdi-spin">mdi-loading</v-icon>
-      <v-icon v-else>mdi-react</v-icon>
+      <v-icon v-else>mdi-printer</v-icon>
     </v-list-item-icon>
     <v-list-item-title>
       {{ $tc('components.print.printReact.downloadNuxtPdfListItem.label') }}
@@ -11,61 +11,11 @@
 </template>
 
 <script>
-import { saveAs } from 'file-saver'
-import slugify from 'slugify'
-
-const RENDER_IN_WORKER = true
+import { generatePdfMixin } from './generatePdfMixin.js'
 
 export default {
   name: 'DownloadReactPdfListItem',
-  props: {
-    config: {
-      type: Object,
-      default: () => {},
-    },
-  },
-  data() {
-    return {
-      loading: false,
-    }
-  },
-  methods: {
-    async generatePdf() {
-      if (this.loading) {
-        return
-      }
-
-      this.loading = true
-
-      // lazy load generatePdf to avoid loading complete react-pdf when showing PDF download button
-      const generatePdfModule = await import('./generatePdf.js')
-
-      const { blob, error } = await generatePdfModule.generatePdf({
-        config: { ...this.config, apiGet: this.api.get.bind(this) },
-        storeData: this.$store.state,
-        translationData: this.$i18n.messages,
-        renderInWorker: RENDER_IN_WORKER,
-      })
-
-      if (error) {
-        this.$emit('error', {
-          label: this.$tc('components.print.printReact.downloadReactPdfListItem.error'),
-          trace: error,
-        })
-        this.loading = false
-        return
-      }
-
-      saveAs(
-        blob,
-        slugify(this.config.documentName, {
-          locale: this.$store.state.lang.language.substr(0, 2),
-        })
-      )
-
-      this.loading = false
-    },
-  },
+  mixins: [generatePdfMixin],
 }
 </script>
