+++ frontend/src/locales/en.json
@@ -138,6 +138,16 @@
       },
       "printPreviewNuxt": {
         "openPreview": "Open preview in new window"
+      },
+      "printReact": {
+        "downloadNuxtPdfListItem": {
+          "label": "Download PDF (layout #2)"
+        }
+      },
+      "printNuxt": {
+        "downloadNuxtPdfListItem": {
+          "label": "Download PDF (layout #1)"
+        }
       }
     },
     "dialog": {
